requirements.txt
selenium==4.15.0
pytest==8.4.2
allure-pytest==2.15.0
pytest-order==1.4.0         # если используете упорядочивание тестов
webdriver-manager==4.11.0   # для автоматического управления драйверами браузеров